Move custom questions to new 'registration' app
This commit creates a new "registration" app within Tabbycat to hold the
logic behind custom questions, allowing for questions to be assigned to
participant models (and possibly more).

Custom questions are now handled in a mixin for feedback forms, and are
available through the API.
