+++ tabbycat/api/fields.py
@@ -7,8 +7,9 @@ from django.utils.encoding import uri_to_iri
 from drf_spectacular.utils import extend_schema_field
 from rest_framework.relations import Hyperlink, HyperlinkedIdentityField, HyperlinkedRelatedField, SlugRelatedField
 from rest_framework.reverse import reverse
-from rest_framework.serializers import CharField, Field, IntegerField
+from rest_framework.serializers import CharField, Field, IntegerField, Serializer, ValidationError
 
+from adjfeedback.models import Question
 from draw.types import DebateSide
 from participants.models import Adjudicator, Speaker, Team
 from venues.models import Venue
@@ -323,3 +324,30 @@ class SideChoiceField(IntegerField):
             if t.pref('teams_in_debate') == 4:
                 return ['og', 'oo', 'cg', 'co'][value]
         return int(value)
+
+
+class AnswerSerializer(Serializer):
+    question = TournamentHyperlinkedRelatedField(
+        view_name='api-question-detail',
+        queryset=Question.objects.all(),
+    )
+    answer = AnyField()
+
+    def validate(self, data):
+        # Convert answer to correct type
+        model = Question.ANSWER_TYPE_CLASSES[data['question'].answer_type]
+        if type(data['answer']) != model.ANSWER_TYPE:
+            raise ValidationError({'answer': 'The answer must be of type %s' % model.ANSWER_TYPE.__name__})
+
+        data['answer'] = model.ANSWER_TYPE(data['answer'])
+
+        option_error = ValidationError({'answer': 'Answer must be in set of options'})
+        if len(data['question'].choices) > 0:
+            if model.ANSWER_TYPE is list and len(set(data['answer']) - set(data['question'].choices)) > 0:
+                raise option_error
+            if data['answer'] not in data['question'].choices:
+                raise option_error
+        if (data['question'].min_value is not None and data['answer'] < data['question'].min_value) or (data['question'].max_value is not None and data['answer'] > data['question'].max_value):
+            raise option_error
+
+        return super().validate(data)
