+++ tabbycat/importer/importers/boots.py
@@ -275,6 +275,7 @@ class BootsTournamentDataImporter(BaseTournamentDataImporter):
             tournament=self.tournament,
             answer_type=self.lookup_feedback_answer_type,
             choices=lambda c: c.split('//'),
+            for_content_type=ContentType.objects.get_for_model(fm.AdjudicatorFeedback),
         )
 
         self._import(f, fm.AdjudicatorFeedbackQuestion, interpreter)
