+++ tabbycat/registration/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class RegistrationConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "registration"
