+++ tabbycat/registration/templates/spaced_choice_widget.html
